move cropper and radio buttons into separate component and add more js tests
